compiler-pm-phase2/requirements.txt
# Python 3.10+
# ============================================

# Base & Utilities
python-dotenv==1.0.0
pydantic==2.4.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# NLP & Parsing
spacy==3.7.2
sqlparse==0.4.4
parso==0.8.3

# Automates & State Machine
transitions==0.9.0

# IA & LLM
openai==1.3.9
langchain==0.1.0
ollama==0.1.0
# For local transformers (optional):
# transformers==4.34.1
# torch==2.1.1

# Dashboard & Visualization
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.3
graphviz==0.20.1
# For interactive diagrams:
pyvis==0.3.2
networkx==3.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
faker==20.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Logging & Monitoring
python-json-logger==2.0.7

# Data Processing
numpy==1.26.2
scipy==1.11.4

# Optional: Production
# gunicorn==21.2.0
# fastapi==0.109.0
# uvicorn==0.24.0
